almost all bugs fixed, need to remove ^ZC, check smashKill, and prepare for delivery

+++ signals.c
@@ -1,11 +1,14 @@
 // signals.c
 #include "signals.h"
+#define _GNU_SOURCE
 
 void handleSIGINT(int sig) {
     // catch SIGINT (Ctrl+C) signal
+    printf("smash: caught CTRL+C\n");
     if(getpid() == globals->smash_pid) {
         longjmp(env_buf, 1); // jump back to the point where setjmp was called
     } else {
+        printf("smash: process %d was killed\n", globals->fgJob->pid); 
         destroyJob(globals->fgJob); // destroy the foreground job
         globals->fgJob = NULL; // clear the foreground job
         exit(0);
@@ -13,22 +16,33 @@ void handleSIGINT(int sig) {
 }
 
 void handleSIGTSTP(int sig) {
+    // catch SIGTSTP (Ctrl+Z) signal
     if(getpid() == globals->smash_pid) {
-        job_t curr_job = globals->fgJob;
-        changeStatus(curr_job, STOPPED); // change the status of the job to STOPPED
-        addExistingJob(curr_job);
-        globals->fgJob = NULL; // clear the foreground job
-        kill(curr_job->pid, SIGSTOP); // send SIGSTOP to the job
+        printf("smash: caught CTRL+Z\n");
+        if (globals->fgJob) { // no command in FOREGROUND
+            job_t curr_job = globals->fgJob;
+            changeStatus(curr_job, STOPPED); // change the status of the job to STOPPED
+            addExistingJob(curr_job);
+            globals->fgJob = NULL; // clear the foreground job
+            kill(curr_job->pid, SIGSTOP); // send SIGSTOP to the job
+            printf("smash: process %d was stopped\n", curr_job->pid); 
+        }
         longjmp(env_buf, 1); // jump back to the point where setjmp was called
     }
     else {
-        printf("ERORORROROROROROROROROROROORORORORORORROORROOROOROROOROROORORORORORROOROROROOROROROOROR");
+        printf("Error: ctrl+z signal received in child process\n");
     }
 }
 
 void setupSignalHandlers() {
     struct sigaction saSIGINT;
     struct sigaction saSIGTSTP;
+
+    // prevent printing ^Z and ^C
+    // struct termios term;
+    // tcgetattr(STDIN_FILENO, &term);
+    // term.c_lflag &= ~ECHOCTL;
+    // tcsetattr(STDIN_FILENO, TCSANOW, &term);
 
     // Setup SIGINT handler
     saSIGINT.sa_handler = handleSIGINT;
